+++ src/main/java/woowacourse/shoppingcart/dto/customer/CustomerRegisterRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.customer;
 
 public class CustomerRegisterRequest {
